port: raid stat-scaling — un-stub MonsterScaling.scaleMonster <weirdgloop@5e313c1>
Port the full ORDER_OF_OPERATIONS chain (ChambersOfXeric/TheatreOfBlood/
TombsOfAmascut/Vardorvis/Phases + the already-ported DefenceReduction) so the
selected monster's stats are scaled by monster.inputs (ToA invocation/path/party,
CoX party + CM, ToB party, Vardorvis current-HP, Araxxor/Yama phases). Add the
missing CoX/ToB id sets to Constants (TOB_MONSTER_IDS, TOB_EM_MONSTER_IDS, OLM_IDS,
COX_MAGIC_IS_DEFENSIVE_IDS, COX_USE_SINGLES_SCALING_IDS, TOA_WARDEN_CORE_EJECTED_IDS,
UNDERWATER_MONSTERS). Fix CalcMath.lerp to mirror upstream's double-precision
divide + single Math.trunc (Math.ts:4-12) — an early long integer division
diverged on the Vardorvis hp-scaling.

+++ src/main/java/org/bossbis/calc/CalcMath.java
@@ -35,9 +35,12 @@ public final strictfp class CalcMath
 
 	public static int lerp(int curr, int srcStart, int srcEnd, int dstStart, int dstEnd)
 	{
-		long srcRange = srcEnd - srcStart;
-		long dstRange = dstEnd - dstStart;
-		long currNorm = curr - srcStart;
+		// Upstream (Math.ts:4-12) does Math.trunc(((currNorm * dstRange) / srcRange) + dstStart) in
+		// double-precision: the division is floating-point and the SINGLE truncation happens AFTER
+		// adding dstStart. Mirror that exactly — an early integer division here diverges (e.g. Vardorvis).
+		double srcRange = srcEnd - srcStart;
+		double dstRange = dstEnd - dstStart;
+		double currNorm = curr - srcStart;
 		return (int) (currNorm * dstRange / srcRange + dstStart);
 	}
 }
